Adds connections between the server and both databases
    - Confirms via basic hardcoded SELECT queries
    - Implements batched CSV writes (performance gains unclear)
